Add DotDict wrapper to support dot access in Transformer 4.28
