+++ pippy/utils.py
@@ -2,7 +2,6 @@
 import os
 import socket
 import logging
-import types
 from typing import List
 
 
@@ -34,7 +33,6 @@ import torch
 import torch.multiprocessing as mp
 import torch.distributed.rpc as rpc
 
-from pippy.PipelineDriver import PipelineDriverBase
 
 PIPPY_VERBOSITY = os.environ.get("PIPPY_VERBOSITY", "OFF")
 
@@ -282,26 +280,3 @@ def run_worker(rank, run_func, args, *extra_args):
         run_func(my_pp_ranks, args, *extra_args)
 
     rpc.shutdown()
-
-
-# This is an experimental utility function that replaces the original model's forward method with PiPPy's PipelineDriver
-# forward method.  It is used to support HuggingFace's `generate()` method, which is defined in a `GenerationMixin`
-# class that `PreTrainedModel` inherits from.  We choose this replacement path instead of writing our own `generate()`
-# method because the `generate()` method would call into many `GenerationMixin` APIs that may be implemented differently
-# by each model.
-def inject_pipeline_forward(
-    model: torch.nn.Module,
-    pipe_driver: PipelineDriverBase,
-):
-    logging.info(
-        f"Inserting PiPPy pipeline forward into model {model._get_name()}"
-    )
-    # Inject pipeline driver as a member object of original model
-    setattr(model, "pippy_pipeline_driver", pipe_driver)
-
-    # Define a new forward method that uses PiPPy's pipeline driver
-    def pippy_forward(self, *args, **kwargs):
-        return self.pippy_pipeline_driver(*args, **kwargs)
-
-    # Replace the forward method in original model
-    setattr(model, "forward", types.MethodType(pippy_forward, model))
